DocumentSettingsTest passed

+++ DocumentModel.BaseTypes/Converters/Percent.XmlSerialization.cs
@@ -40,8 +40,8 @@ public partial struct Percent : IXmlSerializable
       if (!string.IsNullOrEmpty(str))
       {
         str = str.TrimEnd('%');
-        // Parse the percent string to double
-        var parsedValue = decimal.Parse(str.Replace(",", "."), CultureInfo.InvariantCulture);
+        // Parse the percent string to int
+        var parsedValue = int.Parse(str.Replace(",", "."), CultureInfo.InvariantCulture);
 
         // Use Unsafe.AsRef to update the readonly field
         System.Runtime.CompilerServices.Unsafe.AsRef(in value) = parsedValue;
