+++ DocumentModel.BaseTypes/Percent.cs
@@ -8,7 +8,7 @@
 [JsonConverter(typeof(PercentJsonConverter))]
 public readonly partial struct Percent : IConvertible, IEquatable<Percent>, IComparable<Percent>, IComparable<object>
 {
-  private readonly decimal value;
+  private readonly int value;
 
   /// <summary>
   /// Initializes a new instance of the <see cref="Percent"/> struct from a string value.
@@ -20,7 +20,7 @@ public readonly partial struct Percent : IConvertible, IEquatable<Percent>, ICom
   public Percent(string str)
   {
     str = str.TrimEnd('%');
-    value = decimal.Parse(str.Replace(",", "."), CultureInfo.InvariantCulture);
+    value = int.Parse(str.Replace(",", "."), CultureInfo.InvariantCulture);
   }
 
   /// <summary>
@@ -29,7 +29,7 @@ public readonly partial struct Percent : IConvertible, IEquatable<Percent>, ICom
   /// <param name="value">The byte value representing the percentage.</param>
   public Percent(byte value)
   {
-    this.value = (decimal)value;
+    this.value = (int)value;
   }
 
   /// <summary>
@@ -38,7 +38,7 @@ public readonly partial struct Percent : IConvertible, IEquatable<Percent>, ICom
   /// <param name="value">The SByte value representing the percentage.</param>
   public Percent(sbyte value)
   {
-    this.value = (decimal)value;
+    this.value = (int)value;
   }
 
   /// <summary>
@@ -47,7 +47,7 @@ public readonly partial struct Percent : IConvertible, IEquatable<Percent>, ICom
   /// <param name="value">The Int16 value representing the percentage.</param>
   public Percent(Int16 value)
   {
-    this.value = (decimal)value;
+    this.value = (int)value;
   }
 
   /// <summary>
@@ -56,7 +56,7 @@ public readonly partial struct Percent : IConvertible, IEquatable<Percent>, ICom
   /// <param name="value">The Int32 value representing the percentage.</param>
   public Percent(Int32 value)
   {
-    this.value = (decimal)value;
+    this.value = (int)value;
   }
 
 
@@ -66,7 +66,7 @@ public readonly partial struct Percent : IConvertible, IEquatable<Percent>, ICom
   /// <param name="value">The int value representing the percentage.</param>
   public Percent(Int64 value)
   {
-    this.value = (decimal)value;
+    this.value = (int)value;
   }
 
   /// <summary>
@@ -75,7 +75,7 @@ public readonly partial struct Percent : IConvertible, IEquatable<Percent>, ICom
   /// <param name="value">The UInt16 value representing the percentage.</param>
   public Percent(UInt16 value)
   {
-    this.value = (decimal)value;
+    this.value = (int)value;
   }
 
   /// <summary>
@@ -84,7 +84,7 @@ public readonly partial struct Percent : IConvertible, IEquatable<Percent>, ICom
   /// <param name="value">The UInt32 value representing the percentage.</param>
   public Percent(UInt32 value)
   {
-    this.value = (decimal)value;
+    this.value = (int)value;
   }
 
 
@@ -94,7 +94,7 @@ public readonly partial struct Percent : IConvertible, IEquatable<Percent>, ICom
   /// <param name="value">The UInt value representing the percentage.</param>
   public Percent(UInt64 value)
   {
-    this.value = (decimal)value;
+    this.value = (int)value;
   }
 
   /// <summary>
@@ -103,7 +103,7 @@ public readonly partial struct Percent : IConvertible, IEquatable<Percent>, ICom
   /// <param name="value">The Decimal value representing the percentage.</param>
   public Percent(Decimal value)
   {
-    this.value = value;
+    this.value = (int)value;
   }
 
 
@@ -113,7 +113,7 @@ public readonly partial struct Percent : IConvertible, IEquatable<Percent>, ICom
   /// <param name="value">The Single value representing the percentage.</param>
   public Percent(Single value)
   {
-    this.value = (decimal)value;
+    this.value = (int)value;
   }
 
   /// <summary>
@@ -122,7 +122,7 @@ public readonly partial struct Percent : IConvertible, IEquatable<Percent>, ICom
   /// <param name="value">The Double value representing the percentage.</param>
   public Percent(Double value)
   {
-    this.value = (decimal)value;
+    this.value = (int)value;
   }
 
   #region IConvertible Implementation
